fix: code review issues on GetDailyConsumptionBreakdown and allocation responses
- Move ConsumptionGroupDto and ConsumptionDetailDto to Contracts folder
- Replace ex.Message with generic string in GetDailyConsumptionBreakdownHandler catch block
- Change switch default from silent empty list to InvalidOperationException (caught by outer try/catch)
- Add early return when consumptions list is empty to avoid loading full material catalogue
- Strip inline DTO definitions from GetDailyConsumptionBreakdownResponse
- Fix field ordering in GetAllocationsResponse and CreateAllocationResponse to match codebase convention
