+++ backend/src/Anela.Heblo.Application/Features/PackingMaterials/UseCases/GetDailyConsumptionBreakdown/GetDailyConsumptionBreakdownResponse.cs
@@ -1,3 +1,4 @@
+using Anela.Heblo.Application.Features.PackingMaterials.Contracts;
 using Anela.Heblo.Application.Shared;
 
 namespace Anela.Heblo.Application.Features.PackingMaterials.UseCases.GetDailyConsumptionBreakdown;
@@ -8,20 +9,4 @@ public class GetDailyConsumptionBreakdownResponse : BaseResponse
     public DateOnly Date { get; set; }
     public string GroupBy { get; set; } = string.Empty;
     public List<ConsumptionGroupDto> Groups { get; set; } = new();
-}
-
-public class ConsumptionGroupDto
-{
-    public string Key { get; set; } = string.Empty;
-    public string Label { get; set; } = string.Empty;
-    public decimal TotalAmount { get; set; }
-    public int RowCount { get; set; }
-    public List<ConsumptionDetailDto> Details { get; set; } = new();
-}
-
-public class ConsumptionDetailDto
-{
-    public string Key { get; set; } = string.Empty;
-    public string Label { get; set; } = string.Empty;
-    public decimal Amount { get; set; }
 }
